refactor copyright story
